requirements.txt
fastapi
uvicorn
SpeechRecognition
pyaudio
gtts
pygame
gpiozero
python-dotenv
colorama